dying in hell

+++ src/main/java/com/example/prog1demo/units/unit/Archer.java
@@ -6,8 +6,8 @@ import javafx.scene.layout.AnchorPane;
 
 public class Archer extends Generic{
     int cost;
-    int dmg;
-    int hp;
+    double dmg;
+    double hp;
     int movement;
     int initiative;
     int pos_x;
